+++ benchmarks/IfcFilePipelineBenchmark.cs
@@ -1,7 +1,10 @@
 using BenchmarkDotNet.Attributes;
+using BenchmarkDotNet.Configs;
+using BenchmarkDotNet.Jobs;
 
 using Bingosoft.Net.IfcMetadata;
 
+[Config(typeof(IfcFilePipelineBenchmarkConfig))]
 [MemoryDiagnoser]
 public class IfcFilePipelineBenchmark
 {
@@ -42,5 +45,26 @@ public class IfcFilePipelineBenchmark
         }
 
         IfcAccessors.SetTelemetryEnabled(true);
+    }
+}
+
+public sealed class IfcFilePipelineBenchmarkConfig : ManualConfig
+{
+    public IfcFilePipelineBenchmarkConfig()
+    {
+        AddJob(CreateGcJob("WS_Concurrent", gcServer: false, gcConcurrent: true));
+        AddJob(CreateGcJob("WS_NonConcurrent", gcServer: false, gcConcurrent: false));
+        AddJob(CreateGcJob("Server_Concurrent", gcServer: true, gcConcurrent: true));
+    }
+
+    private static Job CreateGcJob(string id, bool gcServer, bool gcConcurrent)
+    {
+        return Job.Default
+            .WithId(id)
+            .WithLaunchCount(1)
+            .WithWarmupCount(3)
+            .WithIterationCount(8)
+            .WithGcServer(gcServer)
+            .WithGcConcurrent(gcConcurrent);
     }
 }